day001

+++ ex5.py
@@ -4,6 +4,6 @@ sell=int(input("enter your selling price: "))
 units=int(input("enter no. of units bought: "))
 
 profit=sell-buy
-final_profit=int(profit*units)
+final_profit=(profit*units)
 print("The profit made on ",stock,"by selling",units,"units of that stock is ",final_profit)
